refactor: レビューしやすいようにディレクトリ構成を変更
