+++ src/components/Board/components/square.tsx
@@ -1,4 +1,4 @@
-import { SquareProps } from "../../../types/type";
+import { SquareProps } from "../../type";
 
 export function Square({
   value,
