+++ src/components/GameStatus/model.ts
